+++ macos/cc/JWMMainView.hh
@@ -5,10 +5,14 @@
 
 namespace jwm {
     extern Key kKeyTable[];
+    extern BOOL kPressAndHoldEnabled;
     void initKeyTable();
 }
 
-@interface JWMMainView : NSView <NSTextInputClient>
+@interface JWMMainView : NSView <NSTextInputClient> {
+    BOOL fInPressAndHold;
+    BOOL fKeyEventsNeeded;
+}
 
 - (JWMMainView*)initWithWindow:(jwm::WindowMac*)initWindow;