+++ shared/cc/Window.cc
@@ -13,14 +13,8 @@ void jwm::Window::dispatch(jobject event) {
         jwm::classes::Consumer::accept(fEnv, fWindow, event);
 }
 
-bool jwm::Window::getRectForMarkedRange(jint selectionStart, jint selectionEnd, jwm::IRect& rect) const {
-    JNILocal<jobject> client(fEnv, fEnv->GetObjectField(fWindow, jwm::classes::Window::kTextInputClient));
-    jwm::classes::Throwable::exceptionThrown(fEnv);
-    if (client.get()) {
-        rect = jwm::classes::TextInputClient::getRectForMarkedRange(fEnv, client.get(), selectionStart, selectionEnd);
-        return true;
-    } else
-        return false;
+jobject jwm::Window::getTextInputClient() const {
+    return fEnv->GetObjectField(fWindow, jwm::classes::Window::kTextInputClient);
 }
 
 extern "C" JNIEXPORT void JNICALL Java_io_github_humbleui_jwm_Window__1nInit