+++ shared/java/EventTextInput.java
@@ -5,7 +5,12 @@ import org.jetbrains.annotations.*;
 
 @Data 
 public class EventTextInput implements Event {
+    /** Text to insert */
     @ApiStatus.Internal public final String _text;
-    // @ApiStatus.Internal public final int _from;
-    // @ApiStatus.Internal public final int _to;
+
+    /** Replacement region start */
+    @ApiStatus.Internal public final int _replacementStart;
+
+    /** Replacement region end */
+    @ApiStatus.Internal public final int _replacementEnd;
 }